Migrated from UIWebView to WKWebView
Since Apple is starting to reject apps using (or referring to) UIWebView I've removed the UIWebView+Extensions class en UIWebViewExtensionTests and created the WKWebView+Extensions and WKWebViewExtensionTests so BFKit-Swift doesn't refer to UIWebView anymore.
